feat:增加自动化任务列表和自动化自动化任务详情的请求

+++ weqa-client-platform-model/src/main/java/com/netease/vcloud/qa/model/VcloudClientAutoAndroidPrefInfoDO.java
@@ -40,7 +40,7 @@ public class VcloudClientAutoAndroidPrefInfoDO implements Serializable {
     private Float instantamperage;
 
     /**
-     * 电量
+     * 功率
      *
      * @mbg.generated
      */
@@ -58,7 +58,7 @@ public class VcloudClientAutoAndroidPrefInfoDO implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer times;
+    private Long times;
 
     /**
      * 绑定taskId
@@ -127,11 +127,11 @@ public class VcloudClientAutoAndroidPrefInfoDO implements Serializable {
         this.level = level;
     }
 
-    public Integer getTimes() {
+    public Long getTimes() {
         return times;
     }
 
-    public void setTimes(Integer times) {
+    public void setTimes(Long times) {
         this.times = times;
     }
